Staging

+++ app/src/main/java/com/nextevent/JavaBeans/CustomRecyclerviewAdapter.java
@@ -3,7 +3,9 @@ package com.nextevent.JavaBeans;
 import android.app.AlertDialog;
 import android.content.Context;
 import android.content.DialogInterface;
+import android.content.SharedPreferences;
 import android.os.Bundle;
+import android.preference.PreferenceManager;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -25,7 +27,8 @@ import androidx.recyclerview.widget.RecyclerView;
 
 /**
  * @author Ghaith Darwish
- * @Last modified: 30/03/2020
+ * @since 07/04/2020
+ * @Last Modified: 08/04/2020
  */
 public class CustomRecyclerviewAdapter extends RecyclerView.Adapter<CustomRecyclerviewAdapter.CustomViewHolder> {
 
@@ -35,18 +38,24 @@ public class CustomRecyclerviewAdapter extends RecyclerView.Adapter<CustomRecycl
     private DatabaseHandler db;
     private boolean showDelete;
 
+    int layout;
+
     public CustomRecyclerviewAdapter(ArrayList<Event> events, Context context, int id, boolean showDelete) {
         this.events = events;
         this.context = context;
         this.id = id;
         db = new DatabaseHandler(context);
         this.showDelete = showDelete;
+
+        // Create SharedPreferences
+        SharedPreferences sharedPreferences = PreferenceManager.getDefaultSharedPreferences(context);
+        layout = sharedPreferences.getInt("LAYOUT", R.layout.event_item_model);
     }
 
     @NonNull
     @Override
     public CustomRecyclerviewAdapter.CustomViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
-        View view = LayoutInflater.from(parent.getContext()).inflate(R.layout.event_item_model, parent, false);
+        View view = LayoutInflater.from(parent.getContext()).inflate(layout, parent, false);
         return new CustomViewHolder(view, id);
     }
 
@@ -54,7 +63,7 @@ public class CustomRecyclerviewAdapter extends RecyclerView.Adapter<CustomRecycl
     public void onBindViewHolder(@NonNull CustomRecyclerviewAdapter.CustomViewHolder holder, int position) {
         Event event = events.get(position);
         holder.title.setText(event.getTitle());
-        holder.date.setText(event.getStart());
+        holder.date.setText(event.getFormattedStartDate());
         holder.location.setText(event.getCountry());
         // Loading images with Picasso
         Picasso.get().load(event.getImage()).placeholder(R.drawable.placeholder).into(holder.image);
@@ -75,6 +84,7 @@ public class CustomRecyclerviewAdapter extends RecyclerView.Adapter<CustomRecycl
 
     /**
      * Setting the ratingBer stars according to the rank
+     *
      * @param rank
      * @param ratingBar
      */
@@ -129,9 +139,9 @@ public class CustomRecyclerviewAdapter extends RecyclerView.Adapter<CustomRecycl
             this.deleteButton = itemView.findViewById(R.id.deleteButton);
             itemView.setOnClickListener(this);
 
-            if (showDelete){
+            if (showDelete) {
                 deleteButton.setVisibility(View.VISIBLE);
-            }else {
+            } else {
                 deleteButton.setVisibility(View.GONE);
             }
 
@@ -148,7 +158,15 @@ public class CustomRecyclerviewAdapter extends RecyclerView.Adapter<CustomRecycl
                                 @Override
                                 public void onClick(DialogInterface dialog, int which) {
                                     Event event = events.get(getAdapterPosition());
-                                    db.deleteEvent(event.getId());
+                                    event.setIsSaved(false);
+
+                                    //Check if the event still needs to be in the database
+                                    if(event.getIsAdded() == 1) {
+                                        db.updateEvent(event);
+                                    } else {
+                                        db.deleteEvent(event.getId());
+                                    }
+
                                     events.remove(event);
                                     notifyItemRemoved(getAdapterPosition());
                                 }
